project.py fixes for uploading hosted images
Two small changes:
(1) When passing in a URL (for hosted image)'self.check_valid_image(path)' doesn't need to run, so we put 'is_hosted' before in the 'or' statement.
(2) When pushing to the API for upload, we had 'self.project_name' which doesn't exists. It should be 'project_name' which is calculated a few lines prior.